Refactor configuration module
