+++ src/python/kiss_icp/config/__init__.py
@@ -1,0 +1,1 @@
+from .parser import KISSConfig, load_config, write_config
